Adapt backend for Vercel serverless
