+++ backend/api/index.ts
@@ -1,0 +1,6 @@
+import serverless from 'serverless-http';
+import { createApp } from '../src/app';
+
+const app = createApp();
+
+export default serverless(app);
