+++ backend/package.json
@@ -21,6 +21,7 @@
     "express": "^4.18.2",
     "express-rate-limit": "^8.2.1",
     "helmet": "^8.1.0",
+    "serverless-http": "^4.0.0",
     "socket.io": "^4.5.4",
     "zod": "^3.22.4"
   },
